+++ benchmarks/source-contentful/bin/site.js
@@ -8,19 +8,6 @@ require("dotenv").config({
   path: `.env.${process.env.NODE_ENV}`,
 })
 
-const siteId = process.env.BENCHMARK_SITE_ID
-const inputDir = path.normalize(
-  `${process.env.BENCHMARK_DATASOURCE_LOCAL_PATH}/data/${siteId}`
-)
-if (!fs.promises.opendir) {
-  console.error(`This script requires Node 12.12+`)
-  process.exit(1)
-}
-if (!fs.existsSync(inputDir)) {
-  console.error(`Could not find datasource directory: ${inputDir}`)
-  process.exit(1)
-}
-
 const locale = `en-US`
 const contentfulConfig = {
   spaceId: process.env.BENCHMARK_CONTENTFUL_SPACE_ID,
@@ -62,6 +49,13 @@ yargs
     desc: `Fix images found by find-broken-images`,
     handler: ({ ids }) => {
       runFixBrokenImages(ids).catch(console.error)
+    },
+  })
+  .command({
+    command: "data-update",
+    desc: `Update random article (to trigger Contentful webhook)`,
+    handler: () => {
+      runDataUpdate().catch(console.error)
     },
   })
   .demandCommand(1)
@@ -110,8 +104,19 @@ function extractEntities(sourceArticle) {
 }
 
 // Node 12.12+
-async function* readSourceArticles(directory) {
-  const dir = await fs.promises.opendir(directory)
+async function* readSourceArticles() {
+  if (!fs.promises.opendir) {
+    console.error(`This command requires Node 12.12+`)
+    process.exit(1)
+  }
+  const inputDir = path.normalize(
+    `${process.env.BENCHMARK_DATASOURCE_LOCAL_PATH}/data/${process.env.BENCHMARK_SITE_ID}`
+  )
+  if (!fs.existsSync(inputDir)) {
+    console.error(`Could not find datasource directory: ${inputDir}`)
+    process.exit(1)
+  }
+  const dir = await fs.promises.opendir(inputDir)
   for await (const dirent of dir) {
     if (dirent.isFile() && /\.json$/.test(dirent.name)) {
       const content = fs.readFileSync(path.join(inputDir, dirent.name))
@@ -254,7 +259,7 @@ async function createEntries({ env, skip = 0 }) {
 
   let total = 0
   let pending = []
-  for await (const sourceArticle of readSourceArticles(inputDir)) {
+  for await (const sourceArticle of readSourceArticles()) {
     if (total++ < skip) {
       continue
     }
@@ -270,7 +275,7 @@ async function createEntries({ env, skip = 0 }) {
 }
 
 async function updateAssets({ env, assetIds }) {
-  for await (const sourceArticle of readSourceArticles(inputDir)) {
+  for await (const sourceArticle of readSourceArticles()) {
     const { asset: assetData } = extractEntities(sourceArticle)
     if (assetIds.has(assetData.sys.id)) {
       try {
@@ -352,4 +357,20 @@ async function runFixBrokenImages(ids) {
   const { env } = await createClient()
   console.log(`Fixing ${chalk.yellow(ids.length)} broken images`)
   await updateAssets({ env, assetIds: new Set(ids) })
+}
+
+async function runDataUpdate() {
+  const { env } = await createClient()
+
+  const entries = await env.getEntries({
+    content_type: 'article',
+    limit: 1
+  })
+
+  const entry = entries.items[0]
+  const title = entry.fields.title[locale]
+  entry.fields.title[locale] = `${title.substring(0, title.lastIndexOf(` `))} ${Date.now()}`
+
+  await entry.update()
+  console.log(`Updated ${chalk.green(entry.sys.id)}`)
 }